# src/main/resources/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tidebound Dota2 Replay Parser API",
    "version": "1.0.0",
    "description": "HTTP interface for parsing Dota 2 replays and managing replay ingestion workflows."
  },
  "servers": [
    {
      "url": "http://localhost:5600",
      "description": "Local parser instance"
    }
  ],
  "paths": {
    "/": {
      "post": {
        "summary": "Parse replay data",
        "description": "Streams a replay payload to the parser for processing. The parser returns the aggregated JSON payload generated by downstream processors.",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsing succeeded. Returns JSON emitted by downstream processors."
          },
          "500": {
            "description": "Parsing failed or replay payload invalid."
          }
        },
        "tags": [
          "Parsing"
        ]
      }
    },
    "/healthz": {
      "get": {
        "summary": "Health check",
        "description": "Returns `ok` when the parser process is healthy.",
        "responses": {
          "200": {
            "description": "Parser is healthy.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "ok"
                }
              }
            }
          }
        },
        "tags": [
          "Operations"
        ]
      }
    },
    "/replay/jobs": {
      "post": {
        "summary": "Submit replay parsing job",
        "description": "Creates an asynchronous job to parse a replay from either a remote URL or a local file path. Returns job metadata so clients can poll for status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "replay_url": {
                    "type": "string",
                    "format": "uri",
                    "description": "Remote HTTP(S) or S3 URL that points to the replay."
                  },
                  "file_path": {
                    "type": "string",
                    "description": "Absolute path to a replay file accessible to the parser."
                  }
                },
                "oneOf": [
                  {
                    "required": [
                      "replay_url"
                    ]
                  },
                  {
                    "required": [
                      "file_path"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted for processing. Returns job metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Exactly one replay source must be provided."
          }
        },
        "tags": [
          "Parsing"
        ]
      }
    },
    "/replay/jobs/{job_id}": {
      "get": {
        "summary": "Get replay parsing job status",
        "description": "Returns the current status of a replay parsing job. When the job succeeds, the parsed result is included in the response.",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayJob"
                }
              }
            }
          },
          "404": {
            "description": "Job not found."
          }
        },
        "tags": [
          "Parsing"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Parsing",
      "description": "Replay ingestion and parsing endpoints."
    },
    {
      "name": "Operations",
      "description": "Health and operational endpoints."
    }
  ],
  "components": {
    "schemas": {
      "ReplayJob": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the job."
          },
          "type": {
            "type": "string",
            "enum": [
              "REMOTE_URL",
              "LOCAL_FILE"
            ],
            "description": "Source type used for the replay."
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "description": "Current lifecycle state of the job."
          },
          "source": {
            "type": "string",
            "description": "Replay location provided when the job was created."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job last changed status."
          },
          "parser_status": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "description": "HTTP-like status returned by the parser process."
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message provided when the job fails."
          },
          "result": {
            "type": "string",
            "nullable": true,
            "description": "Parsed replay payload returned when the job succeeds."
          }
        }
      }
    }
  }
}